I add the generic linked list prom PA7
I also changed the names to reflect something relevant.  to do next, is add the vector of structs, or classes that contains the different types of regular maintaince tasks that can be completed as well as there cost. each customer will have a unique item and cost because there lawns are all unique.  we also need to determin hwo we want the file that holds custoemer properities while the program is not running to look.